refactor/#106: main.ts 정리
