+++ src/app.module.ts
@@ -7,6 +7,7 @@ import { LoggerModule } from "./common/utils/logger/logger.module";
 import { SseModule } from "./sse/sse.module";
 import { RedisModule } from "./auth/redis/redis.module";
 import { ApiModule } from "./api.module";
+import { BootstrapService } from "./bootstrap.service";
 
 @Module({
   imports: [
@@ -20,7 +21,7 @@ import { ApiModule } from "./api.module";
     ApiModule,
   ],
   // isGlobal: ConfigModule을 AppModul 이외의 모듈에서 반복 import할 필요 없는 전역 모듈로 설정.
-  providers: [ConfigService],
+  providers: [ConfigService, BootstrapService],
   // 앵간하면 nestJS에서 권장하는 방법인 ConfigService를 이용하여 환경변수 값을 사용해야함.
 })
 export class AppModule {}
